Prepare for UpdateExpression API
