+++ pynamodb/expressions/util.py
@@ -1,6 +1,11 @@
 import re
+from six import string_types
 
 PATH_SEGMENT_REGEX = re.compile(r'([^\[\]]+)((?:\[\d+\])*)$')
+
+
+def get_path_segments(document_path):
+    return document_path.split('.') if isinstance(document_path, string_types) else list(document_path)
 
 
 def substitute_names(document_path, placeholders):
@@ -15,7 +20,8 @@ def substitute_names(document_path, placeholders):
     that itself is a map element of "foo" (i.e. ['foo', 'bar[0], 'baz']) and an empty placeholders dictionary,
     `substitute_names` will return "#0.#1[0].#2" and placeholders will contain {"foo": "#0", "bar": "#1", "baz": "#2}
     """
-    for idx, segment in enumerate(document_path):
+    path_segments = get_path_segments(document_path)
+    for idx, segment in enumerate(path_segments):
         match = PATH_SEGMENT_REGEX.match(segment)
         if not match:
             raise ValueError('{0} is not a valid document path'.format('.'.join(document_path)))
@@ -25,8 +31,8 @@ def substitute_names(document_path, placeholders):
         else:
             placeholder = '#' + str(len(placeholders))
             placeholders[name] = placeholder
-        document_path[idx] = placeholder + indexes
-    return '.'.join(document_path)
+        path_segments[idx] = placeholder + indexes
+    return '.'.join(path_segments)
 
 
 def get_value_placeholder(value, expression_attribute_values):
