+++ pynamodb/tests/test_table_connection.py
@@ -5,7 +5,7 @@ import six
 from pynamodb.compat import CompatTestCase as TestCase
 from pynamodb.connection import TableConnection
 from pynamodb.constants import DEFAULT_REGION
-from pynamodb.expressions.condition import Path
+from pynamodb.expressions.operand import Path
 from pynamodb.tests.data import DESCRIBE_TABLE_DATA, GET_ITEM_DATA
 from pynamodb.tests.response import HttpOK
